refactor(plugins): update search plugin
