+++ docs/index.html
@@ -10,7 +10,7 @@
   <link rel="stylesheet" href="//unpkg.com/docsify/lib/themes/vue.css">
 </head>
 <body>
-  <nav>
+  <nav data-cloak>
     <a href="#/">En</a>
     <a href="#/zh-cn/">中文</a>
   </nav>
