+++ src/core/global-api.js
@@ -1,7 +1,7 @@
 import * as util from './util'
 import * as dom from './util/dom'
 import * as render from './render/compiler'
-import * as route from './route/util'
+import * as route from './route/hash'
 import { get } from './fetch/ajax'
 import marked from 'marked'
 import prism from 'prismjs'
